burn fix, typo
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@35886 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/ConsoleImpl.cpp
@@ -1737,8 +1737,8 @@ STDMETHODIMP Console::COMGETTER(AttachedPciDevices)(ComSafeArrayOut(IPciDeviceAt
         mBusMgr->listAttachedPciDevices(ComSafeArrayOutArg(aAttachments));
     else
     {
-        com::SafeIfaceArray<IPciDeviceAttachment> result(0);
-        result.detachTo(ComSafeArrayOutArg(aAttachments));            
+        com::SafeIfaceArray<IPciDeviceAttachment> result((size_t)0);
+        result.detachTo(ComSafeArrayOutArg(aAttachments));
     }
 
     return S_OK;
